moved all the crap out of new.rs into their own files
