+++ src/components/i18n_locale.rs
@@ -1,0 +1,28 @@
+use bevy::prelude::*;
+
+#[derive(Component, Default, Reflect, Debug, Clone)]
+#[reflect(Component)]
+pub struct I18nLocale(String);
+impl I18nLocale {
+    pub fn new(locale: impl Into<String>) -> Self {
+        Self(locale.into())
+    }
+
+    pub fn set_locale(&mut self, locale: String) {
+        self.0 = locale;
+    }
+}
+pub trait LocaleExt {
+    fn locale(&self) -> String;
+}
+
+impl LocaleExt for Option<&I18nLocale> {
+    fn locale(&self) -> String {
+        if let Some(internal_locale) = &self {
+            internal_locale.0.to_string()
+        } else {
+            // println!("global {:?} ",  rust_i18n::locale().to_string());
+            rust_i18n::locale().to_string()
+        }
+    }
+}
